fix: always escape keys in exported objects
Previously we were erroneously testing to see if a value was a valid ruby key. This returned true even when the value was not a valid ruby key e.g.

```
'A's Fake Name': {
```

```
Ghtc_Hyuik_1[1]_160x160: {
```
I've change this behavior so that all keys are wrapped in single quotes and, if any single quotes exist in the key string they are escaped e.g.

```
'A\'s Fake Name': {
```

```
'Ghtc_Hyuik_1[1]_160x160': {
```