Forcing a light GC every 100 attachements imported
On some installations we can see upwards of 5GB of attachment data. When we try and load all of this into rails at once, depending on the available machine memory, we can see OOM exceptions which can bring everything down! Yay.

We're injecting a manual GC every 100 attachment loads which will drastically reduce the lifetime of the attachment objects.